# Do not call grpc_init() for per-call-completion-queues created by a C++ sync server