@vjpai 2017-03-29
Beside the minor comment above, this seems LGTM-able. Can you put benchmark results about this into this PR description?