@sreecha 2017-03-30
`gRPC_interop_pull_requests` failure is a known issue: https://github.com/grpc/grpc/issues/9011
(`Bazel Full Build & Tests` suite is not fully implemented yet and we are ignoring those failures)